+++ tests/control/conftest.py
@@ -30,6 +30,7 @@ class BrokenActuationRobot:
 class FakeDynamicsRobot(FakeRobot):
     def __init__(self, actuation_matrix):
         super().__init__(actuation_matrix)
+        self.dynamics_terms_calls = 0
         self._M = jnp.array([[2.0, 0.5], [0.5, 3.0]])
         self._C = jnp.array([[0.0, 0.25], [-0.1, 0.0]])
         self._G = jnp.array([0.3, -0.2])
@@ -50,6 +51,10 @@ class FakeDynamicsRobot(FakeRobot):
 
     def damping_matrix(self, q):
         return self._D
+
+    def dynamics_terms(self, q, qd):
+        self.dynamics_terms_calls += 1
+        return self._M, self._C @ qd, self._G
 
 
 def make_reference(x_des, xd_des, xdd_des=None):